restructuring code and moving file to respective component
