+++ index.html
@@ -11,6 +11,6 @@
     <div id="root">
       <!-- <h1>Hello my name is roopesh</h1> -->
     </div>
-    <script type="module" src="./App.js"></script>
+    <script type="module" src="./src/App.js"></script>
   </body>
 </html>
